Create RSAaaay.py
The script outputs each line in decimal, you have to split up the numbers into their respective values within the ascii range, and convert to ascii. So 103105 would be split to 103 105, which would then be converted to ascii. The proper modifications to the script will soon be added so it does this automatically.